% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{add_scores}
\alias{add_scores}
\title{add_scores}
\usage{
add_scores(df)
}
\arguments{
\item{df}{A QOLPCD dataframe with columns with the following names. The coded
responses to question 1, question 2, etc.: q1, q2, ..., q40. They should be
adjacent to each other and arranged in that order left to right and should
each be coded as integers, 0 - 3 (some questions 0 - 4), higher value corresponding
to higher QOL
Furthermore: version. A string specifying the questionnaire's age group,
either "Ad" (adult), "Asc" (adolescent), "PPx" (parent proxy)}
}
\value{
df amended by columns with the computed scores. Their names are
}
\description{
Calculates the score for all scales in a QOLPCD dataframe and adds them to the
dataframe as separate columns.
}
\examples{
df_qol <- add_scores(df_qol)
colnames(df_qol)
# [...] ...             ...          
# [45]  "q38"           "q39"          
# [47]  "q40"           "physical"     
# [49]  "emotional"     "treatment"    
# [51]  "social"        "role"         
# [53]  "health"        "vitality"     
# [55]  "upper.airways" "lower.airways"
# [57]  "hearing"       "eating"
}
