% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{get_eating}
\alias{get_eating}
\title{get_eating}
\usage{
get_eating(v, version = "PPx")
}
\arguments{
\item{v}{Vector representing one QOLPCD dataset, i.e. a single patient's encoded questionnaire.}

\item{version}{Target age group of the specific questionnaire; "Ad" = adult (default), "Asc" = adolescent, "PPx" = parent proxy}
}
\value{
Vector with responses for the items belonging to the "eating / weight" scale.
}
\description{
Returns vector with responses for the items belonging to the "eating / weight" scale of a given QOLPCD dataset.
}
\examples{
get_eating(df_qol[1,])
}
