% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{get_role}
\alias{get_role}
\title{get_role}
\usage{
get_role(v, version = "Ad")
}
\arguments{
\item{v}{Vector representing one QOLPCD dataset, i.e. a single patient's encoded questionnaire.}

\item{version}{Target age group of the specific questionnaire; "Ad" = adult (default), "Asc" = adolescent, "PPx" = parent proxy}
}
\value{
Vector with responses for the items belonging to the "role" scale.
}
\description{
Returns vector with responses for the items belonging to the "role" scale of a given QOLPCD dataset.
}
\examples{
get_role(df_qol[1,], "Asc")
}
