% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{get_scale}
\alias{get_scale}
\title{get_scale}
\usage{
get_scale(v, version, scale_abbrev)
}
\arguments{
\item{v}{Vector representing one QOLPCD dataset, i.e. a single patient's encoded questionnaire.}

\item{version}{Target age group of the specific questionnaire; "Ad" = adult (default), "Asc" = adolescent, "PPx" = parent proxy}

\item{scale_abbrev}{Abbrevation of the scale's name: P_hysical, E_motional,
T_reatment, S_ocial, R_ole, H_ealth, V_itality, U_pper airways, L_ower airways,
H_earin_g, Ea_ting}
}
\value{
Vector with responses for the items belonging to the specified scale.
}
\description{
Returns vector with responses for the items belonging to the scale of a given QOLPCD dataset specified with its abbreviation.
}
\examples{
get_scale(df_qol[1,], "PPx", "Hg") # returns hearing scale, parent proxy version
}
