% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{get_scale_mean}
\alias{get_scale_mean}
\title{get_scale_mean}
\usage{
get_scale_mean(v, version, scale_abbrev)
}
\arguments{
\item{v}{Vector representing one QOLPCD dataset, i.e. a single patient's encoded questionnaire.}

\item{version}{Target age group of the specific questionnaire; "Ad" = adult (default), "Asc" = adolescent, "PPx" = parent proxy}

\item{scale_abbrev}{Abbrevation of the scale's name: P_hysical, E_motional,
T_reatment, S_ocial, R_ole, H_ealth, V_itality, U_pper airways, L_ower airways,
H_earin_g, Ea_ting}
}
\value{
Mean of the specified scale for the given dataset, considering only
non-missing values, if at least half the values are non-missing. Otherwise
returns NA.
}
\description{
Returns mean of a specific scale for a given QOLPCD dataset.
}
\examples{
get_scale_mean(df_qol[1,], "Ad", "P")
}
