% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{get_upper_respiratory}
\alias{get_upper_respiratory}
\title{get_upper_respiratory}
\usage{
get_upper_respiratory(v, version = "Ad")
}
\arguments{
\item{v}{Vector representing one QOLPCD dataset, i.e. a single patient's encoded questionnaire.}

\item{version}{Target age group of the specific questionnaire; "Ad" = adult (default), "Asc" = adolescent, "PPx" = parent proxy}
}
\value{
Vector with responses for the items belonging to the "upper airways" scale.
}
\description{
Returns vector with responses for the items belonging to the "upper airways" scale of a given QOLPCD dataset.
}
\examples{
get_upper_respiratory(df_qol[1,], "PPx")
}
