% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{impute_and_score}
\alias{impute_and_score}
\title{impute_and_scores}
\usage{
impute_and_score(df)
}
\arguments{
\item{df}{A QOLPCD dataframe with columns with the following names. The coded
responses to question 1, question 2, etc.: q1, q2, ..., q40. They should be
adjacent to each other and arranged in that order left to right and should
each be coded as integers, 0 - 3 (some questions 0 - 4), higher value corresponding
to higher QOL
Furthermore: version. A string specifying the questionnaire's age group,
either "Ad" (adult), "Asc" (adolescent), "PPx" (parent proxy)}
}
\value{
df amended by columns with the computed scores, based on imputed missing
values.
}
\description{
First imputes missing values where appropriate using impute_mean(). Then
calculates the score for all scales in a QOLPCD dataframe and adds them to the
dataframe as separate columns.
}
\examples{
df_qol = read.csv(
    'qolpcd.csv',
    sep=';',
    header=TRUE,
    dec=",",
    na.strings=c(99,88,''),
    encoding='utf-8'
)
df_qol <- impute_and_score(df_qol)
}
