% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{impute_mean}
\alias{impute_mean}
\title{impute_mean}
\usage{
impute_mean(df)
}
\arguments{
\item{df}{A QOLPCD dataframe with columns with the following names. The coded
responses to question 1, question 2, etc.: q1, q2, ..., q40. They should be
adjacent to each other and arranged in that order left to right and should
each be coded as integers, 0 - 3 (some questions 0 - 4), higher value corresponding
to higher QOL
Furthermore: version. A string specifying the questionnaire's age group,
either "Ad" (adult), "Asc" (adolescent), "PPx" (parent proxy)}
}
\value{
df with missing values imputed where appropriate.
}
\description{
Replaces missing values in a QOLPCD dataframe by the mean of the non-missing
values belonging to the same scale in the same dataset. If more than half of
the values that comprise a scale are missing, the missing values are not
imputed.
}
\examples{
impute_mean(df_qol)
}
