% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{include_scale}
\alias{include_scale}
\title{include_scale}
\usage{
include_scale(scale)
}
\arguments{
\item{scale}{Vector comprising responses to questions all belonging to the
same scale of a specific QOLPCD dataset.}
}
\value{
True if at least half the items are non-missing. False otherwise.
}
\description{
Assesses, whether a given scale of a specific QOLPCD dataset contains few
enough missing values to be included in the analyses.
}
\examples{
include_scale(get_physical(df_qol[1,],"PPx"))
}
