% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqolpcd.r
\name{score_scale}
\alias{score_scale}
\title{score_scale}
\usage{
score_scale(v, scale_abbrev, version = "Ad")
}
\arguments{
\item{v}{Vector representing one QOLPCD dataset, i.e. a single patient's encoded questionnaire.}

\item{scale_abbrev}{Abbrevation of the scale's name: P_hysical, E_motional,
T_reatment, S_ocial, R_ole, H_ealth, V_itality, U_pper airways, L_ower airways,
H_earin_g, Ea_ting}

\item{version}{Target age group of the specific questionnaire; "Ad" = adult
(default), "Asc" = adolescent, "PPx" = parent proxy}
}
\value{
The scale's score as a percentage of the maximum achievable score for
that scale and questionnaire version.
}
\description{
Calculates the score of a specific scale for a given QOLPCD dataset.
}
\examples{
score_scale(df_qol[1,], "P", "Ad")
}
